# docs/openapi.yaml
openapi: 3.1.0
info:
  title: HookForge API
  version: 1.2.0
  description: Authenticated, tenant-isolated, at-least-once webhook delivery.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /:
    get:
      security: []
      summary: Service information
      responses:
        "200":
          description: Embedded HookForge operator console
          content:
            text/html:
              schema:
                type: string
  /v1/me:
    get:
      summary: Get the authenticated tenant
      responses:
        "200":
          description: Tenant and API-key identity selected by the Bearer key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Principal"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /healthz:
    get:
      security: []
      summary: Process liveness
      responses:
        "200":
          description: Process is alive
  /readyz:
    get:
      security: []
      summary: Database readiness
      responses:
        "200":
          description: Service is ready
        "503":
          description: Database is unavailable
  /v1/endpoints:
    post:
      summary: Create a webhook endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEndpoint"
      responses:
        "201":
          description: Endpoint created; the signing secret is returned once
          content:
            application/json:
              schema:
                type: object
                required: [endpoint, secret]
                properties:
                  endpoint:
                    $ref: "#/components/schemas/Endpoint"
                  secret:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
    get:
      summary: List endpoints
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Endpoint list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Endpoint"
  /v1/events:
    post:
      summary: Ingest and fan out an event
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvent"
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "202":
          description: New event durably accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResult"
        "200":
          description: Existing event returned for a duplicate key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResult"
        "422":
          $ref: "#/components/responses/ValidationError"
  /v1/events/{id}:
    get:
      summary: Get an event
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/deliveries:
    get:
      summary: List deliveries
      parameters:
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/DeliveryStatus"
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Delivery list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Delivery"
  /v1/deliveries/{id}/replay:
    post:
      summary: Replay a dead-letter delivery
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "204":
          description: Delivery scheduled for replay
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: HookForge API key
      description: The key determines the tenant; clients never submit tenant IDs.
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    Unauthorized:
      description: Missing or invalid tenant API key
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Input validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Principal:
      type: object
      required: [tenant, api_key]
      properties:
        tenant:
          type: object
          required: [id, slug, name]
          properties:
            id:
              type: string
              format: uuid
            slug:
              type: string
            name:
              type: string
        api_key:
          type: object
          required: [id]
          properties:
            id:
              type: string
              format: uuid
    CreateEndpoint:
      type: object
      additionalProperties: false
      required: [name, url]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        url:
          type: string
          format: uri
          description: Must resolve exclusively to public network addresses.
    Endpoint:
      type: object
      required: [id, name, url, enabled, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        url:
          type: string
          format: uri
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateEvent:
      type: object
      additionalProperties: false
      required: [type, payload, endpoint_ids]
      properties:
        type:
          type: string
          minLength: 1
          maxLength: 120
        payload: {}
        endpoint_ids:
          type: array
          minItems: 1
          maxItems: 100
          uniqueItems: true
          items:
            type: string
            format: uuid
    Event:
      type: object
      required: [id, idempotency_key, type, payload, created_at]
      properties:
        id:
          type: string
          format: uuid
        idempotency_key:
          type: string
        type:
          type: string
        payload: {}
        created_at:
          type: string
          format: date-time
    EventResult:
      type: object
      required: [event, duplicate]
      properties:
        event:
          $ref: "#/components/schemas/Event"
        duplicate:
          type: boolean
    DeliveryStatus:
      type: string
      enum: [pending, delivering, succeeded, retrying, dead]
    Delivery:
      type: object
      required:
        - id
        - event_id
        - endpoint_id
        - status
        - attempt_count
        - next_attempt_at
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        endpoint_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/DeliveryStatus"
        attempt_count:
          type: integer
        next_attempt_at:
          type: string
          format: date-time
        last_status_code:
          type: integer
        last_error:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string